requirements.txt
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
confluent-kafka==1.7.0
cryptography==36.0.0
dlib==19.22.1
dnspython==2.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.2
h11==0.12.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
outcome==1.1.0
Pillow==8.4.0
pycparser==2.21
pymongo==3.12.1
pyOpenSSL==21.0.0
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
Werkzeug==2.0.2
wsproto==1.0.0
